fix() split out mat4.mirroring() function and added to index
